refactor: editable src
