+++ packages/editable/src/index.ts
@@ -1,2 +1,14 @@
-export * from "./editable"
-export * from "./use-editable"
+export { Editable, type EditableProps } from "./editable"
+export { EditableInput, type EditableInputProps } from "./editable-input"
+export {
+  EditableTextarea,
+  type EditableTextareaProps,
+} from "./editable-textarea"
+export { EditablePreview, type EditablePreviewProps } from "./editable-preview"
+export {
+  useEditable,
+  type UseEditableProps,
+  type UseEditableReturn,
+} from "./use-editable"
+export { useEditableControls } from "./use-editable-controls"
+export { useEditableState } from "./use-editable-state"
