+++ packages/editable/src/use-editable-controls.tsx
@@ -1,0 +1,21 @@
+import { useEditableContext } from "./editable-context"
+
+/**
+ * React hook use to create controls for the editable component
+ */
+
+export function useEditableControls() {
+  const {
+    isEditing,
+    getEditButtonProps,
+    getCancelButtonProps,
+    getSubmitButtonProps,
+  } = useEditableContext()
+
+  return {
+    isEditing,
+    getEditButtonProps,
+    getCancelButtonProps,
+    getSubmitButtonProps,
+  }
+}
